Added database queries, splash screen, and user persistence
This is a bigger commit than is usually desired - sorry!
Changes:
-Added a splash screen.  This is a view that the app will always start
on, and will display for a minimum of a couple of seconds.  The purpose
is to cover up the time spent checking if there is already stored user
data.  If so, the app proceeds to the home screen, otherwise it goes to
the welcome screen to collect user information.
-Added a database query factory.  Currently this only queries for reps,
not for elections (slight problem - there are no upcoming elections in
the database!)  This can be used in other views to update the rep info,
but be sure to wrap it in a promise, otherwise there will be
asynchronous mumbo jumbo.
-Rep info is now available in $rootScope.repData
-User info is now available in $rootScope.userData
-Home screen autogenerates from data structure
-Contact screen fills from $stateParams
-Info screen fills from $stateParams